# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nutrition Tracker compatibility API
  version: 0.2.0-stage-a
  description: >-
    Repository-derived contract inventory after the approved Stage A
    correctness and security changes. Except for the deliberately empty Vapi
    webhook 401, errors use the ApiError schema.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /food:
    get:
      operationId: getDayLogs
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
      responses:
        '200':
          description: Range logs with array-shaped meals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DayLogResponse' }
...
  /habit:
    get:
      operationId: getAllHabits
      responses:
        '200':
          description: All recurring habits owned by the current user
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
    post:
      operationId: createHabit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HabitRequest' }
      responses:
        '200':
          description: Current persistence-entity JSON shape
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
...
  /habit/{id}/toggle:
    post:
      operationId: toggleHabit
      parameters:
        - $ref: '#/components/parameters/HabitId'
      responses:
        '200': { description: Empty response }
  /habit/{id}:
    delete:
      operationId: deleteHabit
      parameters:
        - $ref: '#/components/parameters/HabitId'
      responses:
        '200': { description: Empty response }
  /habit/voice-result:
    post:
      operationId: submitHabitVoiceResult
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HabitVoiceResult' }
      responses:
        '200':
          description: Updated habit state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HabitWithCompletion' }
  /food/voice-log:
    post:
      security: []
      operationId: handleVapiWebhook
      parameters:
        - name: X-Vapi-Secret
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Provider acknowledgement, including ignored/application-failure events
          content:
            application/json:
              schema:
                type: object
                required: [result]
                properties:
                  result: { type: string, const: logged }
        '401': { description: Empty response when a configured secret does not match }
        '413':
          description: Request body or meal count exceeds the configured limit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...
  /food/voice-log/parse-transcript:
    post:
      operationId: parseVoiceTranscript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transcript: { type: string }
      responses:
        '200':
          description: Transcript processed
          content:
            application/json:
              schema:
                type: object
                required: [status, entriesLogged]
                properties:
                  status: { type: string, const: success }
                  entriesLogged: { type: integer }
        '400': { description: Transcript missing }
        '500': { description: Processing failed }
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    FromDate:
      name: from
      in: query
      required: true
      schema: { type: string, format: date }
    ToDate:
      name: to
      in: query
      required: true
      schema: { type: string, format: date }
...
    HabitId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
...
    ApiError:
      type: object
      required: [status, code, message]
      properties:
        status: { type: integer, example: 400 }
        code: { type: string, example: INVALID_REQUEST }
        message: { type: string, example: Request validation failed }
...
    DayLogEntry:
      type: object
      required: [id, name, quantity, unit, mealType, enrichmentStatus]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        quantity: { type: number }
        unit: { type: string }
        mealType: { type: string }
        nutritionResponse: { type: [string, 'null'] }
        enrichmentStatus:
          type: string
          enum: [pending, in_progress, completed, failed]
        createdAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: [string, 'null'], format: date-time }
    MealEntries:
      type: object
      required: [mealType, entries]
      properties:
        mealType: { type: string }
        entries:
          type: array
          items: { $ref: '#/components/schemas/DayLogEntry' }
    DayLogResponse:
      type: object
      required: [foodLogId, date, meals]
      properties:
        foodLogId: { type: string, format: uuid }
        date: { type: string, format: date }
        meals:
          type: array
          items: { $ref: '#/components/schemas/MealEntries' }
...
    HabitRequest:
      type: object
      properties:
        name: { type: [string, 'null'] }
        repeatDays: { type: [array, 'null'], items: { type: string } }
        reminderTime: { type: [string, 'null'] }
        reminderType: { type: [string, 'null'] }
    HabitWithCompletion:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        repeatDays: { type: array, items: { type: string } }
        reminderTime: { type: string, description: Serialized with pattern hh:mm a }
        reminderType: { type: string }
        completed: { type: boolean }
        status: { type: string }
        completedAt: { type: [string, 'null'] }
        rescheduledTime: { type: [string, 'null'], format: date-time }
    HabitVoiceResult:
      type: object
      properties:
        habitId: { type: [integer, 'null'], format: int64 }
        habitName: { type: [string, 'null'] }
        habitStatus: { type: [string, 'null'] }
        rescheduleMinutes: { type: [integer, 'null'] }
        completedAt: { type: [string, 'null'] }
...